+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Add OTF_PARAMIKO_ULTRA_DEBUG environment variable to enable the hidden `ultra_debug` option for Paramiko.
 - Add some more logging to the SFTP transfers
 - Set `sftp_client` to `None` when closing the SFTP connection
+- Added `OTF_LAZY_LOAD_VARIABLES` environment variable to enable lazy loading of variables.
+- Updated `otflogging` to redact sensitive data from logs
 
 # v24.36.1