+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # v24.36.2
 
-- Add OTF_PARAMIKO_ULTRA_DEBUG environment variable to enable the hidden `ultra_debug` option for Paramiko.
+- Add OTF_PARAMIKO_ULTRA_DEBUG environment variable to enable the hidden `ultra_debug` option for Paramiko (SFTP only).
 - Add some more logging to the SFTP transfers
 - Set `sftp_client` to `None` when closing the SFTP connection
 - Added `OTF_LAZY_LOAD_VARIABLES` environment variable to enable lazy loading of variables.